Add files via upload
Not yet an addon, everything has to be written into the script for it to work. I can see having it run from User Pref, and I can see giving it a string property for name, and another for url, but I'm not sure yet how to handle multiple imports, other than just writing them in manually.